# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Crop": {
        "description": "A crop in the reference catalog, with the agronomic constants the risk\nengine needs. Seeded from a fixture; never client-writable.",
        "properties": {
          "base_temp_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Base temperature for growing-degree-day accumulation.",
            "title": "Base Temp C"
          },
          "category": {
            "$ref": "#/components/schemas/CropCategory"
          },
          "code": {
            "description": "Stable machine key.",
            "examples": [
              "maize"
            ],
            "title": "Code",
            "type": "string"
          },
          "common_diseases": {
            "items": {
              "type": "string"
            },
            "title": "Common Diseases",
            "type": "array"
          },
          "drought_tolerance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DroughtTolerance"
              },
              {
                "type": "null"
              }
            ]
          },
          "gdd_to_maturity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Growing degree days from planting to maturity.",
            "title": "Gdd To Maturity"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "examples": [
              "Maize"
            ],
            "title": "Name",
            "type": "string"
          },
          "optimal_temp_max_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimal Temp Max C"
          },
          "optimal_temp_min_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimal Temp Min C"
          },
          "ph_max": {
            "anyOf": [
              {
                "maximum": 14.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ph Max"
          },
          "ph_min": {
            "anyOf": [
              {
                "maximum": 14.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ph Min"
          },
          "preferred_textures": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Textures",
            "type": "array"
          },
          "scientific_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Zea mays"
            ],
            "title": "Scientific Name"
          },
          "season": {
            "$ref": "#/components/schemas/Season"
          },
          "water_need_mm_season": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Typical seasonal water requirement in mm.",
            "title": "Water Need Mm Season"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "category",
          "season"
        ],
        "title": "Crop",
        "type": "object"
      },
      "CropCategory": {
        "enum": [
          "cereal",
          "legume",
          "oilseed",
          "vegetable",
          "fruit",
          "tuber",
          "fibre",
          "forage",
          "other"
        ],
        "title": "CropCategory",
        "type": "string"
      },
...
      "CropList": {
        "description": "Response for `GET /api/v1/reference/crops`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Crop"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CropList",
        "type": "object"
      },
...
      "DataMode": {
        "description": "Provenance of a data payload. Never omit or guess this.",
        "enum": [
          "live",
          "cached",
          "simulated",
          "unavailable"
        ],
        "title": "DataMode",
        "type": "string"
      },
      "DataSourceMeta": {
        "description": "Provenance attached to every externally-sourced payload.",
        "properties": {
          "fetched_at": {
            "description": "When these values were produced or retrieved.",
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/DataMode",
            "description": "Whether these values are real observations. `simulated` means the data was generated locally and must be labelled as such in the UI."
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable qualifier, e.g. why a provider was degraded.",
            "examples": [
              "SoilGrids timed out; texture-class defaults substituted."
            ],
            "title": "Note"
          },
          "source": {
            "description": "Provider identifier, or 'simulated' when locally generated.",
            "examples": [
              "open-meteo",
              "soilgrids",
              "simulated"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "mode",
          "fetched_at"
        ],
        "title": "DataSourceMeta",
        "type": "object"
      },
...
      "DroughtTolerance": {
        "enum": [
          "low",
          "moderate",
          "high"
        ],
        "title": "DroughtTolerance",
        "type": "string"
      },
      "EnumCatalog": {
        "description": "Every enum the frontend needs to render, with labels.\n\nTyped per-field (rather than a bare dict) so `openapi-typescript` produces a\nreal interface instead of `Record<string, unknown>`.",
        "properties": {
          "advisory_category": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Advisory Category",
            "type": "array"
          },
          "advisory_priority": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Advisory Priority",
            "type": "array"
          },
          "ai_mode": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Ai Mode",
            "type": "array"
          },
          "analysis_status": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Analysis Status",
            "type": "array"
          },
          "crop_category": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Crop Category",
            "type": "array"
          },
          "crop_status": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Crop Status",
            "type": "array"
          },
          "data_mode": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Data Mode",
            "type": "array"
          },
          "drought_tolerance": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Drought Tolerance",
            "type": "array"
          },
          "farming_practice": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Farming Practice",
            "type": "array"
          },
          "growth_stage": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Growth Stage",
            "type": "array"
          },
          "image_analysis_status": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Image Analysis Status",
            "type": "array"
          },
          "irrigation_type": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Irrigation Type",
            "type": "array"
          },
          "risk_level": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Risk Level",
            "type": "array"
          },
          "score_band": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Score Band",
            "type": "array"
          },
          "season": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Season",
            "type": "array"
          },
          "severity": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Severity",
            "type": "array"
          },
          "soil_texture": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "title": "Soil Texture",
            "type": "array"
          }
        },
        "required": [
          "irrigation_type",
          "farming_practice",
          "crop_category",
          "season",
          "growth_stage",
          "crop_status",
          "advisory_category",
          "advisory_priority",
          "severity",
          "analysis_status",
          "ai_mode",
          "image_analysis_status",
          "soil_texture",
          "drought_tolerance",
          "risk_level",
          "score_band",
          "data_mode"
        ],
        "title": "EnumCatalog",
        "type": "object"
      },
      "EnumCatalogResponse": {
        "description": "Wrapper so the catalog can gain metadata without a breaking change.",
        "properties": {
          "enums": {
            "$ref": "#/components/schemas/EnumCatalog",
            "description": "Every enum, with display labels."
          }
        },
        "required": [
          "enums"
        ],
        "title": "EnumCatalogResponse",
        "type": "object"
      },
      "EnumValue": {
        "description": "One selectable option, with display text supplied by the backend.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "label"
        ],
        "title": "EnumValue",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code. Branch on this, never on `message`.",
            "examples": [
              "FARM_NOT_FOUND"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationErrorDetails"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "description": "Structured context, keyed by error code. When `code` is `VALIDATION_ERROR` this is `ValidationErrorDetails` (a `fields` array); otherwise it is a code-specific object, e.g. `{\"farm_id\": \"\u2026\"}` for `FARM_NOT_FOUND`, or empty when there is no extra context.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable explanation. May change without notice.",
            "examples": [
              "Farm 9f8e\u2026 does not exist or is not accessible."
            ],
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "description": "Correlation id, also returned in the X-Request-Id header.",
            "examples": [
              "req_01hzy8k3m2n4p5q6"
            ],
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorField": {
        "description": "One field-level validation failure. Present in `details.fields` on a 422.",
        "properties": {
          "field": {
            "description": "Dotted path to the offending field.",
            "examples": [
              "latitude"
            ],
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "What is wrong with it.",
            "examples": [
              "Input should be less than or equal to 90"
            ],
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "Machine-readable validation failure type.",
            "examples": [
              "less_than_equal"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ErrorField",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The one and only error shape returned by every 4xx/5xx response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Location": {
        "description": "A geocoded place.",
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Kenya"
            ],
            "title": "Country"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "display_name": {
            "description": "Pre-formatted label for a picker row.",
            "examples": [
              "Nakuru, Nakuru County, Kenya"
            ],
            "title": "Display Name",
            "type": "string"
          },
          "elevation_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "examples": [
              "Nakuru"
            ],
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Nakuru County"
            ],
            "title": "Region"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude",
          "display_name"
        ],
        "title": "Location",
        "type": "object"
      },
      "LocationList": {
        "description": "Response for `GET /api/v1/reference/locations`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/DataSourceMeta",
            "description": "Provenance. `simulated` until a real geocoder is wired in."
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "meta"
        ],
        "title": "LocationList",
        "type": "object"
      },
...
      "Season": {
        "enum": [
          "spring",
          "summer",
          "autumn",
          "winter",
          "year_round"
        ],
        "title": "Season",
        "type": "string"
      },
...
      "ValidationErrorDetails": {
        "description": "The shape of `details` when `code` is `VALIDATION_ERROR`.\n\nDocumented as its own model so `ErrorField` is emitted into the OpenAPI components\nand the frontend gets a real type for form-error rendering.",
        "properties": {
          "fields": {
            "description": "Every field that failed validation, one entry per failure.",
            "items": {
              "$ref": "#/components/schemas/ErrorField"
            },
            "title": "Fields",
            "type": "array"
          }
        },
        "required": [
          "fields"
        ],
        "title": "ValidationErrorDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\n**TerraNex** is the central intelligence layer for agricultural decision support.\n\nIt fuses farm, crop, soil, weather and vegetation data with AI reasoning to produce\nfarm health analysis, weather and water risk, disease risk, agricultural advisories,\ncrop and regenerative-agriculture recommendations, and multimodal crop-image diagnosis.\n\n### For frontend developers\n\n* Every endpoint lives under `/api/v1`. Field casing is `snake_case` throughout.\n* Every failure returns the same envelope:\n  `{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}`.\n  **Branch on `code`** \u2014 `message` is for humans and may change.\n* Every response carries `X-Request-Id`; include it when reporting a bug.\n* Generate typed bindings instead of hand-writing them:\n  `npx openapi-typescript ../contracts/openapi.json -o src/api/types.gen.ts`\n",
    "title": "TerraNex API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reference/crops": {
      "get": {
        "description": "The seeded global crop catalog with the agronomic constants used by the risk engine. Read-only reference data, shared by every farm.",
        "operationId": "list_crops_api_v1_reference_crops_get",
        "parameters": [
          {
            "description": "Filter by category.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CropCategory"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category.",
              "title": "Category"
            }
          },
          {
            "description": "Filter by season.",
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Season"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by season.",
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Crop catalog",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/enums": {
      "get": {
        "description": "Every domain enum with display labels. Render dropdowns from this rather than hardcoding values \u2014 new enum members then appear without a frontend change.",
        "operationId": "get_enums_api_v1_reference_enums_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnumCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Enum catalog",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/locations": {
      "get": {
        "description": "Resolves a place name to coordinates for farm registration. The frontend must not call a geocoding provider directly \u2014 the backend is the only integration layer.\n\nCheck `meta.mode`: `simulated` means these are not real geocoder results.",
        "operationId": "search_locations_api_v1_reference_locations_get",
        "parameters": [
          {
            "description": "Place name to search.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Place name to search.",
              "maxLength": 120,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Geocode a place name",
        "tags": [
          "reference"
        ]
      }
    }
  }
}